Empty commit message

+++ src/interfacesComunes/Status.java
@@ -1,11 +1,12 @@
 package interfacesComunes;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.rmi.Remote;
 
 import servidor.TwitterImpl;
 
 public interface Status extends TwitterImpl.ITweet, Serializable, Remote{
 	int getRetweetCount();
-	Long getId();
+	BigInteger getId();
 }
