+++ src/servidor/TwitterEventImpl.java
@@ -1,5 +1,6 @@
 package servidor;
 
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Date;
@@ -44,7 +45,7 @@ public class TwitterEventImpl implements TwitterEvent{
 	public TwitterEventImpl(Long id_source, Long id_target,Status status, String type, Conexion con,User loggedUser) throws SQLException{
 		
 		this.loggedUser=loggedUser;
-		Long id_status=Long.parseLong("0");
+		BigInteger id_status=BigInteger.valueOf(Long.parseLong("0"));
 		this.con=con;
 		this.createdAt=new Date();
 		this.type=type;
@@ -58,7 +59,7 @@ public class TwitterEventImpl implements TwitterEvent{
 		this.status=status;
 		
 		if (this.status==null) //Si el id_status se guarda como 0 es que el Event no afecta a un tweet sino a un User
-			id_status=Long.parseLong("0");
+			id_status=BigInteger.valueOf(Long.parseLong("0"));
 		else id_status=this.status.getId();
 		
 		//Lo anadimos a la base de datos
