+++ src/servidor/UserImpl.java
@@ -1,5 +1,6 @@
 package servidor;
 
+import java.math.BigInteger;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.sql.ResultSet;
@@ -71,7 +72,7 @@ public class UserImpl implements User{
 			this.location = res.getString("location");
 			if(screenName!=null)
 				this.id = res.getLong("id");
-			this.status=new StatusImpl(res.getLong("id_status"),this.con,this.loggedUser);
+			this.status=new StatusImpl(BigInteger.valueOf(res.getLong("id_status")),this.con,this.loggedUser);
 			
 			int c=0;
 			res=con.query("SELECT id_tweet FROM favoritos WHERE id_usuario="+this.id);
